Eliminate some flakiness when not verifying files save
This test was flaking because before the suite runs a bunch of settings
are added which are critical to making the test reliable. Inside the
settings part it opens the editor, edits it and saves the file via
ctrl/cmd+s. This is all fine, but it doesn't verify anything so the
editor may end up closing before ctrl/cmd+s actually gets handled. We
disable the modal in smoke tests since it needs to run headlessly, so
it's difficult to see that the file never actually saves and a dirty
file is closed and discarded.

The fix is to verify settings.json actually does save by changing the
shared Editors.saveOpenedFile mechanism to ensure the dirty indicator
isn't present on the active tab.

Fixes #254893
Part of #246731