tools: update tools of a chat request when they change (#256321)
Makes the `userSelectedTools` passed into the chat agent service
observable such that the tools of the ongoing request update as new
tools come in. The benefit of this versus other (previous) approaches
is that it reflects the enablement/disablement state of the picker
(which is likely to get additional controls in the future) and works
without special tagging.

@roblourens / @DonJayamanne I think with this we can get rid of

https://github.com/microsoft/vscode-copilot-chat/blob/0d6f5516d2fef455340228fd7bc34a4f501725d3/src/extension/tools/vscode-node/toolsService.ts#L118-L122

which was added for Jupyter iirc.

Closes #254684